♻️ refactor: Create select_methods module to assist App class
